Added the constants for the lava generator in the Energy class and added a new class for tanks. Added the enum TransmatricsTileEntity for all tile entities.
Added the lava generator block, on block activated the liquid get transferred from the bucket to the lava generator if the liquid is lava.
Moved the SolarPanel class to the generator package.

Added TileEntityFluidGenerator as base class for generator that use liquids to produce energy.
Added TileEntityLavaGenerator for the tile entity for the lava generator block.

Added ItemStackHelper class for the helper methods for the ItemStack.